offline_benchmark/requirements.txt
edge-tts
pydub
pandas
seaborn
matplotlib
jiwer
psutil
faster-whisper
whisperx
torch
pyannote.audio
